refactor: move staking abi and types to shared catalog
