+++ apps/extension/src/application/trading-copilot/types.ts
@@ -241,15 +241,3 @@ export type VerifySiweSignaturePayload = {
 export type VerifySiweSignatureResponse = {
   token: string;
 };
-
-type ClaimedEvent = {
-  to: string;
-  total: string;
-  bonus: boolean;
-  transactionHash: string;
-};
-
-export type ClaimedEventsResponse = {
-  events: ClaimedEvent[];
-  lastProcessedBlock: string;
-};
