+++ apps/main-landing/src/app/vault/components/stake-tab-content.tsx
@@ -25,15 +25,17 @@ import { useEffect, useState } from 'react';
 import { Spinner } from '@idriss-xyz/ui/spinner';
 import { Checkbox } from '@idriss-xyz/ui/checkbox';
 import { Link } from '@idriss-xyz/ui/link';
-import { TOKEN_TERMS_AND_CONDITIONS_LINK } from '@idriss-xyz/constants';
+import {
+  TOKEN_TERMS_AND_CONDITIONS_LINK,
+  StakingABI,
+  STAKER_ADDRESS,
+} from '@idriss-xyz/constants';
 
 import { ERC20_ABI } from '@/app/creators/donate/constants';
 import { GeoConditionalButton } from '@/components/token-section/components/geo-conditional-button';
 import { TxLoadingModal } from '@/app/claim/components/tx-loading-modal/tx-loading-modal';
-import { IDRISS_TOKEN_ADDRESS } from '@/components/token-section/constants';
 import { formatNumber } from '@/app/claim/components/claim/components/idriss-user-criteria-description';
-
-import { StakingABI, STAKER_ADDRESS } from '../constants';
+import { IDRISS_TOKEN_ADDRESS } from '@/components/token-section/constants';
 
 type FormPayload = {
   amount: number;
