+++ apps/main-landing/src/components/token-section/constants.ts
@@ -1,0 +1,2 @@
+export const IDRISS_TOKEN_ADDRESS =
+  '0x000096630066820566162c94874a776532705231';
