+++ packages/constants/src/index.ts
@@ -21,3 +21,7 @@ export {
   SNAPSHOT_IDRISS_LINK,
   TOKENOMICS_DOCS_LINK,
 } from './links';
+
+export { StakingABI, STAKER_ADDRESS } from './constants';
+
+export type { ClaimedEventsResponse } from './types';
